+++ payroll-backend/src/main/java/edu/mtisw/payrollbackend/repositories/ComprobanteIngresosRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComprobanteIngresosRepository extends JpaRepository<ComprobanteIngresosEntity, Integer>{
+public interface ComprobanteIngresosRepository extends JpaRepository<ComprobanteIngresosEntity, Long>{
 }
