+++ payroll-backend/src/main/java/edu/mtisw/payrollbackend/repositories/UsuarioComprobanteIngresosRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioComprobanteIngresosRepository extends JpaRepository<UsuarioComprobanteIngresosEntity, Integer> {
+public interface UsuarioComprobanteIngresosRepository extends JpaRepository<UsuarioComprobanteIngresosEntity, Long> {
     Optional<UsuarioComprobanteIngresosEntity> findByIdUsuario(Long idUsuario);
 
 }
