+++ payroll-backend/src/main/java/edu/mtisw/payrollbackend/repositories/UsuarioPrestamoRepository.java
@@ -7,8 +7,9 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
-public interface UsuarioPrestamoRepository extends JpaRepository<UsuarioPrestamoEntity, Integer> {
+public interface UsuarioPrestamoRepository extends JpaRepository<UsuarioPrestamoEntity, Long> {
     List<UsuarioPrestamoEntity> findByIdUsuario(Long idUsuario);
 }
